Expressions: Find sprites by filename
A temporary workaround to find character sprites using the character's
filename rather than the character's name. This allows for duplicated
characters to find their sprites.

This was not approved officially by the developers. It's just a hacky
fix until the devs create a proper UI element to override sprite
lookup for a character which shouldn't be too difficult to implement.

Signed-off-by: kingbri <bdashore3@proton.me>